requirements.txt
langchain
openai==0.27.2
python-dotenv==1.0.0
slack_bolt==1.16.4
slack_sdk==3.20.2